feat: Employee register and data.sql, init userTypes

+++ Server/src/main/java/com/teamsantos/easybarber/DTO/UserDTO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 @Data
-@EqualsAndHashCode(callSuper=true)
-public class UserDTO extends BaseResponseDTO{
-    private int             userTypeId;
-    private String          countryMobile;
-    private String          mobile;
+@EqualsAndHashCode(callSuper = true)
+public class UserDTO extends BaseResponseDTO {
+    private Long userTypeId;
+    private String countryMobile;
+    private String mobile;
 }
